+++ src/graphics/color.rs
@@ -55,6 +55,15 @@ impl Color {
 
     /// Shortcut for `Color::rgb(1.0, 1.0, 1.0)`.
     pub const WHITE: Color = Color::rgb(1.0, 1.0, 1.0);
+
+    /// Shortcut for `Color::rgb(1.0, 0.0, 0.0)`.
+    pub const RED: Color = Color::rgb(1.0, 0.0, 0.0);
+
+    /// Shortcut for `Color::rgb(0.0, 1.0, 0.0)`.
+    pub const GREEN: Color = Color::rgb(0.0, 1.0, 0.0);
+
+    /// Shortcut for `Color::rgb(0.0, 0.0, 1.0)`.
+    pub const BLUE: Color = Color::rgb(0.0, 0.0, 1.0);
 }
 
 #[cfg(test)]